# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  discordgo.User:
    properties:
      accent_color:
        description: User's banner color, encoded as an integer representation of
          hexadecimal color code
        type: integer
      avatar:
        description: |-
          The hash of the user's avatar. Use Session.UserAvatar
          to retrieve the avatar itself.
        type: string
      banner:
        description: The hash of the user's banner image.
        type: string
      bot:
        description: Whether the user is a bot.
        type: boolean
      discriminator:
        description: The discriminator of the user (4 numbers after name).
        type: string
      email:
        description: |-
          The email of the user. This is only present when
          the application possesses the email scope for the user.
        type: string
      flags:
        description: |-
          The flags on a user's account.
          Only available when the request is authorized via a Bearer token.
        type: integer
      id:
        description: The ID of the user.
        type: string
      locale:
        description: The user's chosen language option.
        type: string
      mfa_enabled:
        description: Whether the user has multi-factor authentication enabled.
        type: boolean
      premium_type:
        description: |-
          The type of Nitro subscription on a user's account.
          Only available when the request is authorized via a Bearer token.
        type: integer
      public_flags:
        description: |-
          The public flags on a user's account.
          This is a combination of bit masks; the presence of a certain flag can
          be checked by performing a bitwise AND between this int and the flag.
        type: integer
      system:
        description: Whether the user is an Official Discord System user (part of
          the urgent message system).
        type: boolean
      token:
        description: |-
          The token of the user. This is only present for
          the user represented by the current session.
        type: string
      username:
        description: The user's username.
        type: string
      verified:
        description: Whether the user's email is verified.
        type: boolean
    type: object
...
  model.DiscordUserGMStreak:
    properties:
      created_at:
        type: string
      discord_id:
        type: string
      guild_id:
        type: string
      last_streak_date:
        type: string
      streak_count:
        type: integer
      total_count:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.GuildConfigWalletVerificationMessage:
    properties:
      content:
        type: string
      created_at:
        type: string
      discord_message_id:
        type: string
      embedded_message:
        items:
          type: integer
        type: array
      guild_id:
        type: string
      verify_channel_id:
        type: string
    type: object
...
  request.CreateUserRequest:
    properties:
      guild_id:
        type: string
      id:
        type: string
      invited_by:
        type: string
      nickname:
        type: string
      username:
        type: string
    type: object
...
  request.GenerateVerificationRequest:
    properties:
      guild_id:
        type: string
      is_reverify:
        type: boolean
      user_discord_id:
        type: string
    type: object
...
  request.NewGuildConfigWalletVerificationMessageRequest:
    properties:
      content:
        type: string
      created_at:
        type: string
      discord_message_id:
        type: string
      embedded_message:
        items:
          type: integer
        type: array
      guild_id:
        type: string
      verify_channel_id:
        type: string
    type: object
...
  request.VerifyWalletAddressRequest:
    properties:
      code:
        type: string
      signature:
        type: string
      wallet_address:
        type: string
    type: object
...
  response.GenerateVerificationResponse:
    properties:
      code:
        type: string
      status:
        type: string
    type: object
...
  response.GetGuildUserResponse:
    properties:
      guild_id:
        type: string
      invited_by:
        type: string
      nickname:
        type: string
      user_id:
        type: string
    type: object
...
  response.GetMyInfoResponse:
    properties:
      data:
        $ref: '#/definitions/discordgo.User'
    type: object
...
  response.GetUserCurrentGMStreakResponse:
    properties:
      data:
        $ref: '#/definitions/model.DiscordUserGMStreak'
    type: object
  response.GetUserResponse:
    properties:
      data:
        $ref: '#/definitions/response.User'
    type: object
...
  response.NewGuildConfigWalletVerificationMessageResponse:
    properties:
      Data:
        $ref: '#/definitions/model.GuildConfigWalletVerificationMessage'
      status:
        type: string
    type: object
...
  response.ResponseMessage:
    properties:
      message:
        type: string
    type: object
  response.ResponseStatus:
    properties:
      Status:
        type: string
    type: object
...
  response.User:
    properties:
      guild_users:
        items:
          $ref: '#/definitions/response.GetGuildUserResponse'
        type: array
      id:
        type: string
      in_discord_wallet_address:
        type: string
      in_discord_wallet_number:
        type: integer
      username:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a swagger for mochi api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger API
  version: "1.0"
paths:
...
  /profiles/:
    get:
      consumes:
      - application/json
      description: Get user profile
      parameters:
      - description: Guild ID
        in: query
        name: guild_id
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetMyInfoResponse'
      summary: Get user profile
      tags:
      - Profile
...
  /users:
    post:
      consumes:
      - application/json
      description: Create User
      parameters:
      - description: Create user request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseMessage'
      summary: Create User
      tags:
      - User
  /users/{user_id}:
    get:
      consumes:
      - application/json
      description: Get user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetUserResponse'
      summary: Get user
      tags:
      - User
  /users/gmstreak:
    get:
      consumes:
      - application/json
      description: Get user current GM streak
      parameters:
      - description: Discord ID
        in: query
        name: discord_id
        required: true
        type: string
      - description: Guild ID
        in: query
        name: guild_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetUserCurrentGMStreakResponse'
      summary: Get user current GM streak
      tags:
      - User
  /users/me:
    get:
      consumes:
      - application/json
      description: Get user info
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetMyInfoResponse'
      summary: Get user info
      tags:
      - User
  /users/top:
    get:
      consumes:
      - application/json
      description: Get top users
      parameters:
      - description: Guild ID
        in: query
        name: guild_id
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - default: 0
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetMyInfoResponse'
      summary: Get top users
      tags:
      - User
  /verify:
    post:
      consumes:
      - application/json
      description: Verify wallet address
      parameters:
      - description: Verify wallet address request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.VerifyWalletAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseStatus'
      summary: Verify wallet address
      tags:
      - Verification
  /verify/config:
    delete:
      consumes:
      - application/json
      description: Delete guild config wallet verification message
      parameters:
      - description: Guild ID
        in: query
        name: guild_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseStatus'
      summary: Delete guild config wallet verification message
      tags:
      - Verification
    post:
      consumes:
      - application/json
      description: Config wallet verification message
      parameters:
      - description: New guild config wallet verification message request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.NewGuildConfigWalletVerificationMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NewGuildConfigWalletVerificationMessageResponse'
      summary: Config wallet verification message
      tags:
      - Verification
    put:
      consumes:
      - application/json
      description: Update guild config wallet verification message
      parameters:
      - description: Update guild config wallet verification message request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.NewGuildConfigWalletVerificationMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NewGuildConfigWalletVerificationMessageResponse'
      summary: Update guild config wallet verification message
      tags:
      - Verification
  /verify/config/{guild_id}:
    get:
      consumes:
      - application/json
      description: Get guild config wallet verification message
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NewGuildConfigWalletVerificationMessageResponse'
      summary: Get guild config wallet verification message
      tags:
      - Verification
  /verify/generate:
    post:
      consumes:
      - application/json
      description: Generate verification
      parameters:
      - description: Generate verification request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.GenerateVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GenerateVerificationResponse'
      summary: Generate verification
      tags:
      - Verification
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: access_token
    type: apiKey
swagger: "2.0"